Rename top-level package to 'link'
