+++ link/adapters/gateway.py
@@ -5,10 +5,10 @@ from collections import defaultdict
 from itertools import groupby
 from typing import Iterable
 
-from dj_link.domain.custom_types import Identifier
-from dj_link.domain.link import Link, create_link
-from dj_link.domain.state import Commands, Components, Processes, Update
-from dj_link.service.gateway import LinkGateway
+from link.domain.custom_types import Identifier
+from link.domain.link import Link, create_link
+from link.domain.state import Commands, Components, Processes, Update
+from link.service.gateway import LinkGateway
 
 from .custom_types import PrimaryKey
 from .facade import DJAssignments, DJLinkFacade, DJProcess
