+++ pyproject.toml
@@ -37,7 +37,7 @@ disable = ["cyclic-import", "logging-fstring-interpolation", "duplicate-code"]
 [tool.mypy]
 python_version = "3.8"
 files = [
-    "dj_link", 
+    "link", 
     "tests",
     "stubs",
 ]
@@ -78,7 +78,7 @@ ignore_missing_imports = true
 
 [tool.coverage.run]
 branch = true
-source = ["dj_link"]
+source = ["link"]
 
 [tool.coverage.report]
 exclude_lines = ["if TYPE_CHECKING:"]
@@ -126,7 +126,7 @@ markers = [
 line-length = 120
 select = ["E", "F", "I", "D", "PL", "PT"]
 ignore = ["F821"]
-include = ["dj_link/*.py", "tests/*.py"]
+include = ["link/*.py", "tests/*.py"]
 
 [tool.ruff.per-file-ignores]
 "tests/*.py" = ["D", "PL"]
