+++ tests/functional/test_deleting.py
@@ -1,6 +1,6 @@
 import datajoint as dj
 
-from dj_link import link
+from link import link
 
 
 def test_deleting(
